Moving to rust
made a cargo repo
move every modules from anilistpy/mutation , anilistpy/user ,
anilistpy/search into anilistpy

modified setup.py and cargo.toml
