+++ Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "anilistpy"
+version = "0.1.0"
+authors = ["0kalekale <kalekale.anon@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+ureq = "2.0.1"
+
+[dependencies.pyo3]
+version = "0.13.2"
+features = ["extension-module"]
+
+[lib]
+name = "anilistpy"
+crate-type = ["cdylib"]
