+++ anilistpy/__init__.py
@@ -1,19 +1,22 @@
+from .anilistpy import *
 from anilistpy.anime import Anime
 from anilistpy.manga import Manga
 from anilistpy.char import Character
 from anilistpy.staff import Staff
 from anilistpy.studio import Studio
-from anilistpy.search.manga import mangaSearch
-from anilistpy.search.anime import animeSearch
-from anilistpy.search.char import charSearch
-from anilistpy.search.staff import searchStaff
-from anilistpy.search.studio import searchStudio
+
+from anilistpy.sr_manga import mangaSearch
+from anilistpy.sr_anime import animeSearch
+from anilistpy.sr_char import charSearch
+from anilistpy.sr_staff import searchStaff
+from anilistpy.sr_studio import searchStudio
 from anilistpy.utils import videoLink
 
-from anilistpy.mutation.media import Media
-from anilistpy.mutation.user import User
+from anilistpy.mt_media import Media
+from anilistpy.mt_user import User
 
-from anilistpy.user.medialist import MediaList 
+from anilistpy.usr_medialist import MediaList 
+
 '''
     easy to use wrapper for anilist apiv2
     Copyright (C) 2020  0kalekale
