--- anilistpy/mutation/__init__.py
@@ -1,19 +1,0 @@
-from anilistpy.mutation.media import Media
-from anilistpy.mutation.user import User
-'''
-    easy to use wrapper for anilist apiv2
-    Copyright (C) 2020  0kalekale
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <https://www.gnu.org/licenses/>.
-'''
