--- anilistpy/search/__init__.py
@@ -1,23 +1,0 @@
-from anilistpy.search.anime import animeSearch
-from anilistpy.search.manga import mangaSearch
-from anilistpy.search.char import charSearch
-from anilistpy.search.staff import searchStaff
-from anilistpy.search.studio import searchStudio
-
-'''
-    easy to use wrapper for anilist apiv2
-    Copyright (C) 2020  0kalekale
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <https://www.gnu.org/licenses/>.
-'''
