+++ anilistpy/sr_manga.py
